Review helpers

+++ app/helpers/admin/adminsite_application_helper.rb
@@ -1,41 +1,5 @@
 module Admin::AdminsiteApplicationHelper
 
-  def column_of_attr(search_attr)
-    return if search_attr.blank?
-    resource_class.columns.each{|c| return c if c.name == search_attr.to_s.downcase }
-    nil
-  end
-
-  def input_type_of_column(column)
-    case
-    when column.sql_type.match(/\Acharacter varying/)
-      return :string
-    when column.sql_type.match(/\Atimestamp/)
-      return :date
-    when column.sql_type.match(/\integer/)
-      return :number
-    else
-      column.sql_type.try(:to_sym)
-    end
-  end
-
-  def ransack_predicate_input_type(input_type)
-    case input_type
-    when :boolean
-      return [:eq]
-    when :text
-      return [:eq, :cont]
-    when :string
-      return [:eq, :cont]
-    when :date
-      return [:lteq, :gteq]
-    when :number
-      return [:eq, :lteq, :gteq]
-    else
-      [:eq]
-    end
-  end
-
   def format_response_value(value, add_td_wrappers = true)
     value = value.url if defined?(PictureUploader) && value.is_a?(PictureUploader)
     response = ''
@@ -57,95 +21,18 @@ module Admin::AdminsiteApplicationHelper
     value.is_a?(String) && value.match(/\A[\/]|\Ahttp[s]*:/)
   end
 
-  def image_extensions
-    %w(.png .gif .jpg .tif)
-  end
-
   def is_image?(path)
     image_extensions.include?( File.extname(path).split('?').first.try(:downcase) )
+  end
+
+  def image_extensions
+    %w(.png .gif .jpg .tif)
   end
 
   def error_messages_for(obj)
     return if obj.errors.blank?
     msgs = obj.errors.full_messages.collect{|msg| "<li>#{ h msg }</li>" }
     raw ['<ul>', msgs, '</ul>'].flatten.join
-  end
-
-  def recognize_path(path)
-    return {} if path.try(:strip).blank?
-    begin
-      return Adminsite::Engine.routes.recognize_path(path) # '/admin/profiles'
-      return Rails.application.routes.recognize_path(path)
-    rescue Exception => e
-    end
-  end
-
-  def current_admin_menu
-    @current_admin_menu ||= params[:admin_menu]
-  end
-
-  def menu_item(label, url, child_controller_names = [], klasses = nil, method = nil, admin_menu = label)
-    result = ''
-    child_menus = []
-
-    child_controller_names_authorized = child_controller_names.select do |child_controller_name|
-      controller_class_name = "#{child_controller_name}_controller".classify
-      controller_class = eval("defined?(Adminsite::" + "#{controller_class_name}) ? Adminsite::#{controller_class_name} : Adminsite::Admin::#{controller_class_name}".classify)
-      can?(:read, controller_class.new.authorize_resource_class)
-    end
-
-    if current_url?(url, label) || ( child_controller_active?(child_controller_names) && current_admin_menu == admin_menu)
-      child_controller_names_authorized.each do |child_controller_name|
-        child_menus << content_menu_item(child_controller_name, admin_menu, nil, nil )
-      end
-    end
-    child_menus = child_menus.compact
-
-    if child_controller_names_authorized.count > 0
-      html_options = {method: method}
-      html_options[:title] = current_adminsite_admin_user.email if klasses == 'log_out'
-      link = link_to(label, "#{url}?admin_menu=#{admin_menu}", html_options)
-      result = raw "<li class='#{html_classes(url, nil, klasses, admin_menu, label )}'>#{link}</li>"
-
-      content_for(:content_menu, child_menus.join("\n").html_safe )
-    end
-    result
-  end
-
-  def content_menu_label(url, controller_name)
-    menu_controller = recognize_path(url)[:controller]
-    return controller_name.titlecase if menu_controller.blank?
-    eval("#{menu_controller}_controller".classify).content_menu_label
-  end
-
-  def content_menu_item(controller_name, admin_menu, klasses, method )
-    if controller_name != controller_name.pluralize
-      url = eval("admin_#{controller_name}_index_path")
-    else
-      url = eval("admin_#{controller_name}_path")
-    end
-    link = link_to(content_menu_label(url, controller_name), "#{url}?admin_menu=#{admin_menu}", method: method, )
-    raw "<li class='#{html_classes(url, controller_name, klasses, admin_menu )}'>#{link}</li>"
-  end
-
-  def current_url?(url, label = '')
-    if label.present?
-      return request.fullpath == "#{url}?admin_menu=#{label}"
-    else
-      return request.fullpath == url
-    end
-  end
-
-  def child_controller_active?(child_controller_names)
-    child_controller_names.include?(controller.controller_name)
-  end
-
-  def html_classes(url, controller_name, klasses = nil, admin_menu = '', label = '')
-    result = []
-    result |= [klasses].flatten if klasses.present?
-    result |= ['current'] if current_url?(url, admin_menu)
-    result |= ['active']  if current_admin_menu == label || controller_name == controller.controller_name
-    result.join(' ')
   end
 
   def label_resource
@@ -169,32 +56,47 @@ module Admin::AdminsiteApplicationHelper
   end
 
   def link_to_show(resource)
-    link_to image_tag('adminsite/admin/magnifier.png', :size => '16x16'), admin_resource_path(resource.id), target: '_blank'
+    link_to image_tag('adminsite/admin/magnifier.png', :size => '16x16'), admin_resource_path(resource.id), target: '_blank' if can?(:read, resource)
   end
 
   def link_to_edit(resource)
-    link_to image_tag('adminsite/admin/pencil.png', :size => '16x16'), admin_resource_path(resource.id, :edit)
+    link_to image_tag('adminsite/admin/pencil.png', :size => '16x16'), admin_resource_path(resource.id, :edit) if can?(:edit, resource)
   end
 
   def link_to_destroy(resource)
-    link_to image_tag('adminsite/admin/cross.png', :size => '16x16'), admin_resource_path(resource.id), data: { :confirm => 'Are you sure?'} , :method => :delete
+    link_to image_tag('adminsite/admin/cross.png', :size => '16x16'), admin_resource_path(resource.id), data: { :confirm => 'Are you sure?'} , :method => :delete if can?(:destroy, resource)
   end
 
   def display_resource_value(resource, attr, add_td_wrappers = true)
     value = nil
     attr.to_s.split('.').each{|a| value = (value || resource).send(a) }
-    value = display_referenced_resource(value) if value.is_a?(ActiveRecord::Base)
-    if value.is_a?(ActiveRecord::Associations::CollectionProxy)
+    value = display_referenced_resource(value, false) if value.is_a?(ActiveRecord::Base)
+    if value.is_a?(ActiveRecord::Associations::CollectionProxy) or value.is_a?(Array)
       value = value.collect do |r|
-        label = Adminsite::AdminConfig::Base.admin_config_of_class(r.class, nil, current_adminsite_admin_user).label_attribute(r)
-        display_resource_value(r, label, false)
+        if r.is_a?(ActiveRecord::Base)
+          display_referenced_resource(r, false)
+        else
+          r
+        end
       end.join(', ')
     end
     format_response_value(value, add_td_wrappers).html_safe
   end
 
-  def display_referenced_resource(resource)
-    link_to resource.title, send("edit_admin_#{resource.class.name.underscore.gsub('/','_')}_path", resource.id, admin_menu: params[:admin_menu])
+  def display_referenced_resource(resource, add_td_wrappers = true)
+    label_attr = Adminsite::AdminConfig::Base.admin_config_of_class(resource.class, nil, current_adminsite_admin_user).label_attribute(resource)
+    label = display_resource_value(resource, label_attr, add_td_wrappers)
+    begin
+      if can?(:edit, resource)
+        link_to label, send("edit_admin_#{resource.class.name.underscore.gsub('/','_')}_path", resource.id, admin_menu: params[:admin_menu])
+      elsif can?(:read, resource)
+        link_to label, send("admin_#{resource.class.name.underscore.gsub('/','_')}_path", resource.id, admin_menu: params[:admin_menu])
+      else
+        label
+      end
+    rescue
+      return label
+    end
   end
 
   def publish_form_input(form, form_input)
